test(stripe): split plugin shape and metadata helper smoke into own files
Pulls the type-level expectTypeOf describe out of stripe.test.ts into
plugin.test.ts so the runtime test file no longer carries plugin-shape
smoke. Internal-field protection and getter helpers move into
metadata.test.ts next to the prototype pollution guard so all
metadata tests live in one place. stripe.test.ts loses ~140 lines and
the now-unused Auth/StripePlugin/expectTypeOf imports.
