+++ packages/stripe/test/metadata.test.ts
@@ -61,3 +61,47 @@ describe("stripe metadata prototype pollution guard", () => {
 		expect(result.customerType).toBe("user");
 	});
 });
+
+describe("stripe metadata helpers", () => {
+	it("customerMetadata.set protects internal fields", () => {
+		const result = customerMetadata.set(
+			{ userId: "real", customerType: "user" },
+			{ userId: "fake", custom: "value" },
+		);
+		expect(result.userId).toBe("real");
+		expect(result.customerType).toBe("user");
+		expect(result.custom).toBe("value");
+	});
+
+	it("customerMetadata.get extracts typed fields", () => {
+		const result = customerMetadata.get({
+			userId: "u1",
+			customerType: "organization",
+			extra: "ignored",
+		});
+		expect(result.userId).toBe("u1");
+		expect(result.customerType).toBe("organization");
+		expect(result).not.toHaveProperty("extra");
+	});
+
+	it("subscriptionMetadata.set protects internal fields", () => {
+		const result = subscriptionMetadata.set(
+			{ userId: "u1", subscriptionId: "s1", referenceId: "r1" },
+			{ subscriptionId: "fake" },
+		);
+		expect(result.subscriptionId).toBe("s1");
+	});
+
+	it("subscriptionMetadata.get extracts typed fields", () => {
+		const result = subscriptionMetadata.get({
+			userId: "u1",
+			subscriptionId: "s1",
+			referenceId: "r1",
+			extra: "ignored",
+		});
+		expect(result.userId).toBe("u1");
+		expect(result.subscriptionId).toBe("s1");
+		expect(result.referenceId).toBe("r1");
+		expect(result).not.toHaveProperty("extra");
+	});
+});
