+++ packages/stripe/test/stripe.test.ts
@@ -1,147 +1,13 @@
 import { runWithEndpointContext } from "@better-auth/core/context";
-import type { Auth, User } from "better-auth";
+import type { User } from "better-auth";
 import { organization } from "better-auth/plugins/organization";
 import { getTestInstance } from "better-auth/test";
 import type Stripe from "stripe";
-import { assert, describe, expect, expectTypeOf, it, vi } from "vitest";
-import type { StripePlugin } from "../src";
+import { assert, describe, expect, vi } from "vitest";
 import { stripe } from "../src";
 import { stripeClient } from "../src/client";
-import { customerMetadata, subscriptionMetadata } from "../src/metadata";
 import type { StripeOptions, Subscription } from "../src/types";
 import { test } from "./_fixtures";
-
-describe("stripe type", () => {
-	it("should api endpoint exists", () => {
-		type Plugins = [
-			StripePlugin<{
-				stripeClient: Stripe;
-				stripeWebhookSecret: string;
-				subscription: {
-					enabled: false;
-				};
-			}>,
-		];
-		type MyAuth = Auth<{
-			plugins: Plugins;
-		}>;
-		expectTypeOf<MyAuth["api"]["stripeWebhook"]>().toBeFunction();
-	});
-
-	it("should have subscription endpoints", () => {
-		type Plugins = [
-			StripePlugin<{
-				stripeClient: Stripe;
-				stripeWebhookSecret: string;
-				subscription: {
-					enabled: true;
-					plans: [];
-				};
-			}>,
-		];
-		type MyAuth = Auth<{
-			plugins: Plugins;
-		}>;
-		expectTypeOf<MyAuth["api"]["stripeWebhook"]>().toBeFunction();
-		expectTypeOf<MyAuth["api"]["subscriptionSuccess"]>().toBeFunction();
-		expectTypeOf<MyAuth["api"]["listActiveSubscriptions"]>().toBeFunction();
-		expectTypeOf<MyAuth["api"]["cancelSubscription"]>().toBeFunction();
-		expectTypeOf<MyAuth["api"]["restoreSubscription"]>().toBeFunction();
-		expectTypeOf<MyAuth["api"]["upgradeSubscription"]>().toBeFunction();
-		expectTypeOf<MyAuth["api"]["createBillingPortal"]>().toBeFunction();
-	});
-
-	test("should infer plugin schema fields on user type", async ({
-		stripeMock,
-		memory,
-		stripeOptions,
-	}) => {
-		const { auth } = await getTestInstance({
-			plugins: [
-				stripe({
-					stripeClient: {} as Stripe,
-					stripeWebhookSecret: "test",
-				}),
-			],
-		});
-		expectTypeOf<
-			(typeof auth)["$Infer"]["Session"]["user"]["stripeCustomerId"]
-		>().toEqualTypeOf<string | null | undefined>();
-	});
-
-	test("should infer plugin schema fields alongside additional user fields", async ({
-		stripeMock,
-		memory,
-		stripeOptions,
-	}) => {
-		const { auth } = await getTestInstance({
-			plugins: [
-				stripe({
-					stripeClient: {} as Stripe,
-					stripeWebhookSecret: "test",
-				}),
-			],
-			user: {
-				additionalFields: {
-					customField: {
-						type: "string",
-						required: false,
-					},
-				},
-			},
-		});
-		expectTypeOf<
-			(typeof auth)["$Infer"]["Session"]["user"]["stripeCustomerId"]
-		>().toEqualTypeOf<string | null | undefined>();
-		expectTypeOf<
-			(typeof auth)["$Infer"]["Session"]["user"]["customField"]
-		>().toEqualTypeOf<string | null | undefined>();
-	});
-});
-
-describe("stripe - metadata helpers", () => {
-	it("customerMetadata.set protects internal fields", () => {
-		const result = customerMetadata.set(
-			{ userId: "real", customerType: "user" },
-			{ userId: "fake", custom: "value" },
-		);
-		expect(result.userId).toBe("real");
-		expect(result.customerType).toBe("user");
-		expect(result.custom).toBe("value");
-	});
-
-	it("customerMetadata.get extracts typed fields", () => {
-		const result = customerMetadata.get({
-			userId: "u1",
-			customerType: "organization",
-			extra: "ignored",
-		});
-		expect(result.userId).toBe("u1");
-		expect(result.customerType).toBe("organization");
-		expect(result).not.toHaveProperty("extra");
-	});
-
-	it("subscriptionMetadata.set protects internal fields", () => {
-		const result = subscriptionMetadata.set(
-			{ userId: "u1", subscriptionId: "s1", referenceId: "r1" },
-			{ subscriptionId: "fake" },
-		);
-		expect(result.subscriptionId).toBe("s1");
-	});
-
-	it("subscriptionMetadata.get extracts typed fields", () => {
-		const result = subscriptionMetadata.get({
-			userId: "u1",
-			subscriptionId: "s1",
-			referenceId: "r1",
-			extra: "ignored",
-		});
-		expect(result.userId).toBe("u1");
-		expect(result.subscriptionId).toBe("s1");
-		expect(result.referenceId).toBe("r1");
-		expect(result).not.toHaveProperty("extra");
-	});
-});
 
 describe("stripe", () => {
 	const testUser = {
